requirements.txt
joblib>=1.4.0
numpy>=1.26.0
Pillow>=10.0.0
scikit-learn>=1.4.0
streamlit>=1.40.0
tensorflow>=2.15.0
